project restructure and modularisation of components
