+++ frontend/src/App.jsx
@@ -1,7 +1,7 @@
 import { Routes, Route } from 'react-router-dom'
 import Home from './pages/Home'
-import Navbar from './components/Navbar'
-import Footer from './components/Footer'
+import Navbar from './components/ui/Navbar'
+import Footer from './components/ui/Footer'
 import Register from './pages/Register'
 import Login from './pages/Login'
 import Dashboard from './pages/Dashboard'
