+++ frontend/src/pages/Dashboard.jsx
@@ -4,36 +4,30 @@ import UserDetailsCard from '../components/UserDetailsCard'
 import UserWeightCard from '../components/UserWeightsCard'
 import WorkoutFeed from '../components/WorkoutFeed'
 import WorkoutCreationForm from '../components/forms/WorkoutCreationForm'
-import WorkoutDetails from '../components/WorkoutDetails' // Import WorkoutDetails
+import WorkoutDetailsPreview from '../components/WorkoutDetailsPreview'
+import useWorkoutDetails from '../hooks/useWorkoutDetails'
+import axios from 'axios'
 
 const Dashboard = () => {
     const { accessToken } = useContext(AuthContext)
     const [user, setUser] = useState(null)
     const [weights, setWeights] = useState([])
     const [workoutId, setWorkoutId] = useState(null) // Store selected workoutId
-    const [isModalOpen, setIsModalOpen] = useState(false) // Modal state
-    const [isCreating, setIsCreating] = useState(false) // Track creating state
-    const [isSubmitted, setIsSubmitted] = useState(false) // Track form submission state
+    const [isModalOpen, setIsModalOpen] = useState(false)
+    const [isCreating, setIsCreating] = useState(false)
+    const [isSubmitted, setIsSubmitted] = useState(false)
 
+    // ✅ Fetch User Details and Weight History using Axios
     useEffect(() => {
         const fetchUserDetails = async () => {
             try {
-                const response = await fetch(
+                const response = await axios.get(
                     `${process.env.REACT_APP_API_BASE_URL}/users/me/`,
-                    {
-                        headers: {
-                            Authorization: `Bearer ${accessToken}`,
-                        },
-                    }
+                    { headers: { Authorization: `Bearer ${accessToken}` } }
                 )
 
-                if (!response.ok) {
-                    throw new Error('❌ Failed to fetch user details')
-                }
-
-                const data = await response.json()
-                setUser(data.user)
-                setWeights(data.weights.slice(-10)) // Last 10 weigh-ins
+                setUser(response.data.user)
+                setWeights(response.data.weights.slice(-10)) // Last 10 weigh-ins
             } catch (error) {
                 console.error('❌ Error fetching user details:', error)
             }
@@ -42,17 +36,18 @@ const Dashboard = () => {
         fetchUserDetails()
     }, [accessToken])
 
+    // ✅ Handles opening/closing workout creation modal
     const handleOpenModal = () => setIsModalOpen(true)
     const handleCloseModal = () => {
         setIsModalOpen(false)
-        setIsSubmitted(false) // Reset after closing modal
+        setIsSubmitted(false)
     }
 
     return (
         <div className="pt-24 pb-20 flex flex-col items-center min-h-screen bg-[#8B0000] text-white">
             {user && <UserDetailsCard user={user} />}
 
-            {/* Create Workout Button (only visible when form is closed) */}
+            {/* ✅ Create Workout Button (hidden when form is open) */}
             {!isModalOpen && !isCreating && !isSubmitted && !workoutId && (
                 <button
                     onClick={handleOpenModal}
@@ -62,14 +57,14 @@ const Dashboard = () => {
                 </button>
             )}
 
-            {/* Show Creating Workout Card when submitting */}
+            {/* ✅ Show "Creating Workout" Status */}
             {isCreating && (
                 <div className="w-full max-w-md bg-yellow-400 text-black text-center p-3 mb-4 mt-6 rounded-lg shadow-md">
                     Creating workout...
                 </div>
             )}
 
-            {/* Modal for Workout Creation */}
+            {/* ✅ Workout Creation Modal */}
             {isModalOpen && (
                 <WorkoutCreationForm
                     onClose={handleCloseModal}
@@ -78,12 +73,12 @@ const Dashboard = () => {
                 />
             )}
 
-            {/* Display WorkoutDetails if a workout is selected */}
+            {/* ✅ Display WorkoutDetails if a workout is selected */}
             {workoutId && (
                 <div className="mt-6 w-full max-w-6xl">
-                    <WorkoutDetails workoutId={workoutId} />
+                    <WorkoutDetailsPreview workoutId={workoutId} />
                     <button
-                        onClick={() => setWorkoutId(null)} // Reset to return to workout list
+                        onClick={() => setWorkoutId(null)}
                         className="bg-red-500 text-black px-4 py-2 rounded mt-4 hover:bg-red-400 transition"
                     >
                         Back to Workout Feed
@@ -91,7 +86,7 @@ const Dashboard = () => {
                 </div>
             )}
 
-            {/* Show Workout Feed & Weight Chart only if no workout is selected */}
+            {/* ✅ Show Workout Feed & Weight Chart only when no workout is selected */}
             {!workoutId && (
                 <div className="w-full max-w-6xl grid grid-cols-1 md:grid-cols-2 gap-6 mt-6">
                     {weights.length > 0 && <UserWeightCard weights={weights} />}
